Harrison - 3465 - #120 - KJON/SWYN: Move PredictedEphemerisTracker to its own file
